fix the duplication route methods
